code/requirements.txt
django
psycopg2-binary
pillow
django-silk
django-ninja
pydantic
django-ninja-simple-jwt